+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/implementation/servicebus/AzureServiceBusAutoConfigurationTests.java
@@ -8,6 +8,7 @@ import com.azure.messaging.servicebus.ServiceBusClientBuilder;
 import com.azure.messaging.servicebus.models.ServiceBusReceiveMode;
 import com.azure.spring.cloud.autoconfigure.implementation.AbstractAzureServiceConfigurationTests;
 import com.azure.spring.cloud.autoconfigure.implementation.context.properties.AzureGlobalProperties;
+import com.azure.spring.cloud.autoconfigure.implementation.servicebus.properties.AzureServiceBusConnectionDetails;
 import com.azure.spring.cloud.autoconfigure.implementation.servicebus.properties.AzureServiceBusProperties;
 import com.azure.spring.cloud.core.properties.profile.AzureEnvironmentProperties;
 import com.azure.spring.cloud.core.provider.AzureProfileOptionsProvider;
@@ -81,7 +82,7 @@ class AzureServiceBusAutoConfigurationTests extends AbstractAzureServiceConfigur
         azureProperties.getCredential().setClientSecret("azure-client-secret");
         azureProperties.getRetry().getExponential().setBaseDelay(Duration.ofSeconds(2));
 
-        this.contextRunner
+        this.getMinimalContextRunner()
             .withBean("azureProperties", AzureGlobalProperties.class, () -> azureProperties)
             .withPropertyValues(
                 "spring.cloud.azure.servicebus.credential.client-id=servicebus-client-id",
@@ -104,7 +105,7 @@ class AzureServiceBusAutoConfigurationTests extends AbstractAzureServiceConfigur
         AzureGlobalProperties azureProperties = new AzureGlobalProperties();
         azureProperties.getProfile().setCloudType(AzureProfileOptionsProvider.CloudType.AZURE_US_GOVERNMENT);
 
-        this.contextRunner
+        this.getMinimalContextRunner()
                 .withBean("azureProperties", AzureGlobalProperties.class, () -> azureProperties)
                 .withPropertyValues(
                         "spring.cloud.azure.servicebus.domain-name=servicebus.chinacloudapi.cn"
@@ -120,7 +121,7 @@ class AzureServiceBusAutoConfigurationTests extends AbstractAzureServiceConfigur
 
     @Test
     void configureAmqpTransportTypeShouldApply() {
-        this.contextRunner
+        this.getMinimalContextRunner()
             .withBean("azureProperties", AzureGlobalProperties.class, AzureGlobalProperties::new)
             .withPropertyValues("spring.cloud.azure.servicebus.client.transport-type=AmqpWebSockets")
             .run(context -> {
@@ -132,7 +133,7 @@ class AzureServiceBusAutoConfigurationTests extends AbstractAzureServiceConfigur
 
     @Test
     void configureRetryShouldApply() {
-        this.contextRunner
+        this.getMinimalContextRunner()
             .withBean("azureProperties", AzureGlobalProperties.class, AzureGlobalProperties::new)
             .withPropertyValues(
                 "spring.cloud.azure.servicebus.retry.mode=fixed",
@@ -295,6 +296,22 @@ class AzureServiceBusAutoConfigurationTests extends AbstractAzureServiceConfigur
             .run(context -> {
                 assertThat(context).doesNotHaveBean(AzureServiceBusProcessorClientConfiguration.class);
                 assertThat(context).hasSingleBean(AzureServiceBusConsumerClientConfiguration.class);
+            });
+    }
+
+    @Test
+    void connectionDetailsHasHigherPriority() {
+        String connectionString = String.format(CONNECTION_STRING_FORMAT, "property-namespace");
+        this.contextRunner
+            .withPropertyValues(
+                "spring.cloud.azure.servicebus.connection-string=" + connectionString
+            )
+            .withBean(AzureGlobalProperties.class, AzureGlobalProperties::new)
+            .withBean(AzureServiceBusConnectionDetails.class, CustomAzureServiceBusConnectionDetails::new)
+            .run(context -> {
+                assertThat(context).hasSingleBean(AzureServiceBusProperties.class);
+                AzureServiceBusProperties properties = context.getBean(AzureServiceBusProperties.class);
+                assertEquals(CustomAzureServiceBusConnectionDetails.CONNECTION_STRING, properties.getConnectionString());
             });
     }
 }